change sesseionId from uint64 to uint32, for align problem

+++ services/distributeddataservice/libs/distributeddb/common/include/runtime_context.h
@@ -125,7 +125,7 @@ public:
     virtual int NotifyUserChanged() const = 0;
 
     // Generate global sessionId in current process
-    virtual uint64_t GenerateSessionId() = 0;
+    virtual uint32_t GenerateSessionId() = 0;
 protected:
     RuntimeContext() = default;
     virtual ~RuntimeContext() {}
