+++ services/distributeddataservice/libs/distributeddb/common/src/runtime_context_impl.h
@@ -114,7 +114,7 @@ public:
     // Notify TIME_CHANGE_EVENT.
     int NotifyUserChanged() const override;
 
-    uint64_t GenerateSessionId() override;
+    uint32_t GenerateSessionId() override;
 private:
     static constexpr int MAX_TP_THREADS = 10;  // max threads of the task pool.
     static constexpr int MIN_TP_THREADS = 1;   // min threads of the task pool.
@@ -170,7 +170,7 @@ private:
     mutable std::mutex userChangeMonitorLock_;
     std::unique_ptr<UserChangeMonitor> userChangeMonitor_;
 
-    std::atomic<uint64_t> currentSessionId_;
+    std::atomic<uint32_t> currentSessionId_;
 };
 } // namespace DistributedDB
