+++ services/distributeddataservice/libs/distributeddb/storage/src/iconnection.cpp
@@ -30,7 +30,7 @@ uint64_t IConnection::GetConnectionId()
     std::lock_guard<std::mutex> autoLock(connectionIdLock_);
     // check again here, may be generated after get lock
     if (connectionId_ == 0) {
-        connectionId_ = RuntimeContext::GetInstance()->GenerateSessionId();
+        connectionId_ = static_cast<uint64_t>(RuntimeContext::GetInstance()->GenerateSessionId());
     }
     return connectionId_;
 }
